requirements.txt
torch>=2.1.0
fastapi>=0.110.0
uvicorn>=0.29.0
python-multipart>=0.0.9
pydantic>=2.7.0
pyyaml>=6.0.1
sentencepiece>=0.2.0
pyarrow>=16.0.0
psutil>=5.9.8
wandb>=0.17.0
pytest>=8.2.0
httpx>=0.27.0
bitsandbytes>=0.43.0
